+++ lib/auto_stories.dart
@@ -1,7 +1,6 @@
 /// Auto generate data and widget stories preview once file changed.
 library;
 
-export 'src/generate/annotations.dart';
+export 'src/analysis/annotations.dart';
 export 'src/state.dart';
-export 'src/system.dart';
 export 'src/utils.dart';