+++ lib/src/analysis.dart
@@ -1,0 +1,2 @@
+export 'analysis/annotations.dart';
+export 'analysis/package.dart';