+++ lib/src/system.dart
@@ -1,1 +1,0 @@
-export 'system/package.dart';