+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `feat`: handler for bidirectional data update in widget tree.
 - `feat`: react as reactive and flexible data handler outside widget tree.
 - `feat`: dart package abstraction and common utilities.
+- `feat`: utilities about file and folder copy.
 
 - `perf`: remove unnecessary linter rules.
 - `docs`: more readable docs in readme.